Dev mode, descriptive JSON, 1 job, start timer on page load
Set pages and JS to "dev mode", added a fake config grabber function
to allow for more varying and longer tests.

Changed the description of the json, what used to be status is now
running.

Grabber will only allow one job.

The runningRefreshTimer on the status page will start on page load
if the server has an active job running. A background timer will check
the server periodically to see if it's running.