Closes #3 Adds custom device types
Device types and methods can now be defined without having
to recompile the binary! Types are defined in config/device-types.conf
where a type string, method (for example ssh/telnet on a cisco type),
the script filename to be ran for the type, and a list of arguments
in the order that the script is expecting them. This will allow
for easy expansion of new types by the end user. Any script can
be launched so long as it resides in the scripts directory.

The default types are cisco ssh, cisco telnet, and juniper ssh.
This update removes the need for a TFTP server. The logs are saved
to the latest directory instead of a TFTP directory. The Cisco
scripts now use log_file instead of copying to a tftp server.
All this allows for a wider range of devices and appliances to be
used with this application.