Good about of design done, updated libraries, start on new API
The rough design for the dashboard, device list, and device management
pages are done. They're not functional but look pretty. Bootstrap and
jQuery have been updated. Added a pie chart JS library and d3 to make
a pretty neat pie chart on the dashboard. It's a big library but
it's the only one I could customize to my liking.

The api is started but nothing has been tested yet. All requests are
routed to the apiHandler which then routes to individual subsystems.
Which themselves route to specific functions. Every request
responses with a standard JSON object similar to the response
object in my Dandelion project. It has an error message, error code,
path name, and data properties.