Devices create and edit endpoints same, wire up frontend
I decided to combine the create and edit devices api endpoints
to a single save endpoint. A device id of -1 denotes a new device.
Anything else will be treated as an edit to an existing device.

The frontend is starting to get wired up. It doesn't do any api
calls yet, but it looks really nice.