Wired live device status on dashboard
The Dashboard now gets actual data from the application to render
a pie chart and table of down devices. Right now the devices link
to their configs (which don't exist yet) but I may change that. The
new api is "/devices/status" and returns the number of total, up,
down, and unknown devices along with information for devices that
are not up.