Finished core to poller communication
The core has a job index where a function sends a job to the poller and
receives a channel that will contain the result of the job when it's
returned from the poller.

The Poller has a job queue where new requests are added and then
dispatched by a separate goroutine. The job goroutines then write
the results back to the core. Jobs and responses are connected
via a job ID sent to the poller.