Removed external poller, centralized configuration
I decided having the poller as an external executable was too much
of a hassle for what Inca will be doing. Its jobs aren't so intensive
that they need a totally separate process to run. Goroutines will work
just fine. Plus, the new poller package is much easier to work with.

The configuration has been centralized in the common package. Inca
now accepts an argument for the location of the configuration file.
The argument is -c.