Created poller, device configs are grabbed or pinged
The poller is now (mostly) finished apart from some minor tweaks
and inevitable bug fixes. It will check if the device is supported
for a config grab based on its connection profile. If it's not
supported it will simply ping the devices using the system pinger.
A pinger in Go would require being ran as root. Since this is
being designed as a Linux only system I see no harm is using the
native pinger.

The device status is updated as well as a config entry is added to
the database. The next step is viewing configs.