Removed settings handler, don't delete old configs, new logging
Logging is a little simpler. Logs are written to files within the same
directory instead of individual folders.

The settings handler is removed. It showed too much.

Old configs will no be deleted unless a new config was created
successfully. This prevents configs being unavailable when a device
isn't reachable which is usually when you need them the most.